+++ src/main/java/kamitesque/common/entities/EntityItemPersistent.java
@@ -1,10 +1,16 @@
 package kamitesque.common.entities;
 
-import kamitesque.events.front.PersistentItemEvents;
+import kamitesque.util.PersistentUtils;
+import net.minecraft.entity.Entity;
+import net.minecraft.entity.EntityLiving;
 import net.minecraft.entity.item.EntityItem;
+import net.minecraft.entity.monster.IMob;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.item.ItemStack;
+import net.minecraft.util.math.AxisAlignedBB;
 import net.minecraft.world.World;
+
+import java.util.List;
 
 public class EntityItemPersistent extends EntityItem {
 
@@ -12,17 +18,10 @@ public class EntityItemPersistent extends EntityItem {
 
     public EntityItemPersistent(World worldIn, double x, double p_i1710_3_, double y, ItemStack p_i1710_5_) {
         super(worldIn, x, p_i1710_3_, y, p_i1710_5_);
-        init();
-    }
 
-    public EntityItemPersistent(World worldIn) {
-        super(worldIn);
-        init();
-    }
-
-    private void init() {
         this.setEntityInvulnerable(true);
         this.setNoDespawn();
+        this.setPickupDelay(20);
 
         this.rotationYaw = (float)(Math.random() * (double)360.0F);
         this.motionX = (double)((float)(Math.random() * (double)0.2F - (double)0.1F));
@@ -30,19 +29,18 @@ public class EntityItemPersistent extends EntityItem {
         this.motionZ = (double)((float)(Math.random() * (double)0.2F - (double)0.1F));
     }
 
+    public EntityItemPersistent(World worldIn) {
+        super(worldIn);
+    }
+
     protected void dealFireDamage(int amount) {}
 
     public void onCollideWithPlayer(EntityPlayer entityIn) {
 
-        if (entityIn.world.isRemote) {
-            return;
-        }
-
         if (this.getItem().getTagCompound() != null && this.getItem().getTagCompound().hasKey("kamitesque.persistent")) {
             if (!this.getItem().getTagCompound().getUniqueId("kamitesque.persistent.owner").equals(entityIn.getUniqueID())) {
 
-                PersistentItemEvents.punish(entityIn);
-
+                PersistentUtils.punish(entityIn);
                 return;
             }
         }
@@ -50,6 +48,40 @@ public class EntityItemPersistent extends EntityItem {
         super.onCollideWithPlayer(entityIn);
     }
 
+    public void onUpdate() {
+
+            if (this.ticksExisted > 1) {
+                if (this.motionY > (double)0.0F) {
+                    this.motionY *= (double)0.9F;
+                }
+
+                this.motionY += (double)0.04F;
+            }
+
+            if (this.ticksExisted > 10) {
+                this.motionX *= 0.5F;
+                this.motionZ *= 0.5F;
+            }
+
+            AxisAlignedBB box = this.getEntityBoundingBox().grow(2);
+
+            List<Entity> entities = this.world.getEntitiesWithinAABBExcludingEntity(this, box);
+
+            if (!entities.isEmpty()) {
+                for (Entity entity : entities) {
+                    if (entity instanceof EntityLiving && !(entity instanceof EntityPlayer)) {
+                        ((EntityLiving)entity).setCanPickUpLoot(false);
+
+                        if (entity instanceof IMob) {
+                            entity.setFire(5);
+                        }
+
+                    }
+                }
+            }
+            super.onUpdate();
     }
+
+}