+++ src/main/java/kamitesque/init/KTSounds.java
@@ -7,13 +7,18 @@ import net.minecraftforge.registries.IForgeRegistry;
 public class KTSounds {
 
     public static SoundEvent banish;
-
     public static SoundEvent blazing;
+
+    public static SoundEvent stamping;
+    public static SoundEvent persistence;
 
     public static void initSounds(IForgeRegistry<SoundEvent> iForgeRegistry) {
 
         iForgeRegistry.register(banish = new SoundEventKTBase("banish"));
         iForgeRegistry.register(blazing = new SoundEventKTBase("blazing"));
+
+        iForgeRegistry.register(stamping = new SoundEventKTBase("stamping"));
+        iForgeRegistry.register(persistence = new SoundEventKTBase("persistence"));
 
     }
 }