+++ src/main/java/kamitesque/network/proxy/ClientProxy.java
@@ -4,9 +4,11 @@ import kamitesque.client.renderer.tiles.RenderIchorflameNitor;
 import kamitesque.client.renderer.tiles.RenderRootCrystal;
 import kamitesque.common.tiles.TileIchorflameNitor;
 import kamitesque.common.tiles.TileRootCrystal;
+import kamitesque.init.KTRender;
 import kamitesque.root.Main;
 import net.minecraftforge.client.model.obj.OBJLoader;
 import net.minecraftforge.fml.client.registry.ClientRegistry;
+import net.minecraftforge.fml.common.event.FMLInitializationEvent;
 import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
 
 public class ClientProxy extends CommonProxy {
@@ -17,6 +19,11 @@ public class ClientProxy extends CommonProxy {
 
         ClientRegistry.bindTileEntitySpecialRenderer(TileIchorflameNitor.class, new RenderIchorflameNitor());
         ClientRegistry.bindTileEntitySpecialRenderer(TileRootCrystal.class, new RenderRootCrystal());
+    }
+
+    @Override
+    public void init(FMLInitializationEvent event) {
+        KTRender.initEntityRender();
     }
 
 }