+++ src/main/java/kamitesque/root/Main.java
@@ -28,7 +28,7 @@ import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
 public class Main {
     public static final String MODID = "kamitesque";
     public static final String NAME = "Kamitesque";
-    public static final String VERSION = "0.0.8-ALPHA";
+    public static final String VERSION = "0.0.9-ALPHA";
 
     @SidedProxy(clientSide = "kamitesque.network.proxy.ClientProxy", serverSide = "kamitesque.network.proxy.CommonProxy")
     public static CommonProxy proxy;
@@ -40,7 +40,6 @@ public class Main {
     public void preInit(FMLPreInitializationEvent event) {
 
         KTNetwork.preInitPackets();
-
         KTTiles.preInitTiles();
         KTEntities.preInitEntities();
 
@@ -52,6 +51,8 @@ public class Main {
 
         KTResearch.initResearch();
         KTResearch.initScans();
+
+        proxy.init(event);
     }
 
     @EventHandler